requirements.txt
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
scipy==1.13.1
sqlalchemy==2.0.30
jupyter==1.0.0
notebook==7.2.0
ipykernel==6.29.4
openpyxl==3.1.2

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Configuration
pyyaml>=6.0
